--- 404.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" style="width:100%;height:100%;">
-  <head>
-    <meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>Not found - Browsercraft</title>
-		<script defer data-domain="browsercraft.cheerpj.com" src="https://plausible.leaningtech.com/js/script.js"></script>
-		<link rel="stylesheet" href="style.css">
-
-		<meta name="description" content="Open-source port of Minecraft to the web. Play Minecraft in your browser!">
-		<meta property="og:title" content="Browsercraft">
-		<meta property="og:description" content="Open-source port of Minecraft to the web. Play Minecraft in your browser!">
-		<meta property="og:url" content="https://browsercraft.cheerpj.com">
-	</head>
-  <body>
-    <main>
-      <h1>
-        Not found
-      </h1>
-      <p>
-        This page was not found. Try going back to the <a href="/">homepage</a>.
-      </p>
-    </main>
-  </body>
-</html>
