+++ README.md
@@ -1,5 +1,37 @@
 # Browsercraft
 
-This is a proof of concept of Minecraft running unmodified in the browser, using [CheerpJ](https://labs.leaningtech.com/cheerpj). 
+## What is it
+
+Thanks to [Cheerpj](https://cheerpj.com), this project runs an alpha version of Minecraft (1.2.5) on the browser, without the requirement of having the JVM locally installed on the computer.
 
 See [the website](https://browsercraft.cheerpj.com) for a live demo and more information.
+
+## Notes
+Since this is a demo the playing time is restricted to 3 minutes; see [here](https://www.minecraft.net) for the full game.
+
+This is not an official Minecraft product. It is not approved by or associated with Mojang or Microsoft, furthermore, the original JARs of the game are downloaded from the ufficial website and used untouched.
+
+For questions and support about CheerpJ and Browsercraf join our [Discord](https://discord.gg/7xXW6NAdHT) server.
+
+
+## Execute locally
+
+First clone the project:
+
+```
+git clone git@github.com:leaningtech/browsercraft.git
+```
+
+Install the dependencies:
+
+```
+npm install
+```
+
+Run in development mode:
+
+```
+npm run dev
+```
+
+The site will be available at `localhost:5173`
