--- _headers
@@ -1,2 +1,0 @@
-/libraries/*
-  Content-Type: application/octet-stream
