--- embed.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" style="width:100%;height:100%;overflow:hidden">
-  <head>
-    <meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>Browsercraft</title>
-		<script src="https://cjrtnc.leaningtech.com/3_20250303_606/cj3loader.js"></script>
-		<script defer data-domain="browsercraft.cheerpj.com" src="https://plausible.leaningtech.com/js/script.js"></script>
-    <link rel="stylesheet" href="style.css">
-	</head>
-  <body>
-  <div class="container" style="width: 100%; height: 100%">
-    <div class="intro">
-      <p>
-        This is a proof-of-concept demo of Minecraft 1.2.5 running unmodified in the browser.
-      </p>
-      <p>
-        Clicking the button below will download the client from mojang.com.
-        By clicking it, you agree to the <a href="https://www.minecraft.net/eula">Minecraft EULA</a>.
-      </p>
-      <button>Play!</button>
-      <div class="disclaimer">
-        This is not an official Minecraft product. It is not approved by or associated with Mojang or Microsoft.
-      </div>
-    </div>
-  <progress style="display: none"></progress>
-  <div class="display"></div>
-</div>
-    <script type="module">
-			import MinecraftClient from "./minecraft-web.js";
-			await cheerpjInit({
-				version: 8,
-				javaProperties: ["java.library.path=/app/libraries/"],
-				libraries: {"libGL.so.1": "/app/libraries/gl4es.wasm", "libEGL.so.1": "/app/libraries/libegl.wasm", "libGLESv2.so.1": "/app/libraries/libgles.wasm"},
-				enableX11:true
-			});
-			const mc = new MinecraftClient();
-    </script>
-  </body>
-</html>
