+++ src/app.css
@@ -1,0 +1,178 @@
+html, body {
+	height: 100%;
+	width: 100%;
+	margin: 0;
+	padding: 0;
+
+	font-family: system-ui, sans-serif;
+}
+
+@font-face {
+	font-family: "Minecrafter";
+	src: url("/fonts/minecrafter/Minecrafter.Reg.woff2"), sans-serif;
+}
+
+header {
+	background: #171615;
+	color: white;
+	padding: 1rem;
+
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+}
+
+header h1 {
+	color: #c6b8b4;
+	font: 48px Minecrafter;
+	text-shadow: 0px 5px 4px black;
+	transform: perspective(24px) translateZ(0) rotate3d(1, 0, 0, 2deg);
+}
+
+main {
+	margin: 0 auto;
+	padding: 1rem;
+	max-width: 60ch;
+	width: 100%;
+	box-sizing: border-box;
+	line-height: 1.4;
+}
+
+main h2 {
+	font-size: 1.5rem;
+	margin: 1rem 0 0.5rem;
+}
+
+main li {
+	margin: 0.5rem 0;
+}
+
+@media (max-width: 900px) {
+	minecraft-client {
+		width: 100%;
+	}
+}
+
+.pagecontrols {
+	margin: 0.5rem 0;
+}
+
+.pagecontrols > * {
+	appearance: none;
+	background: transparent;
+	border: 0;
+	cursor: pointer;
+
+	width: 24px;
+	height: 24px;
+
+	margin-right: 0.5rem;
+}
+
+.pagecontrols svg {
+	stroke: #c6b8b4;
+}
+
+.pagecontrols > *:hover svg {
+	stroke: white;
+}
+
+.pagecontrols > #share {
+	display: none;
+}
+
+.container {
+	display: inline-block;
+	aspect-ratio: 4 / 3;
+
+	background: black;
+	color: #eee;
+	color-scheme: dark;
+
+	width: 800px;
+	height: 500px;
+	position: relative;
+}
+
+canvas {
+	width: inherit;
+	height: inherit;
+}
+
+.display {
+	width: 800px;
+	height: 500px;
+	position: relative;
+	inset: 0;
+	display: none;
+}
+
+.intro {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	height: 100%;
+}
+
+.intro > p {
+	max-width: 60ch;
+}
+
+.timeout {
+	position: absolute;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	display: none;
+	background-color: rgba(0, 0, 0, 0.8);
+}
+
+.timeout > p {
+	max-width: 60ch;
+}
+
+.timeout-timer {
+	position: absolute;
+	font-size: 4rem;
+	margin: 0;
+	top: 0;
+	right: 0;
+	display: none;
+}
+
+.timeout-timer > p {
+	margin-top: 0;
+	margin-bottom: 0;
+	margin-right: 1rem;
+	color: white;
+	text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.8),
+							 -1px -1px 2px rgba(0, 0, 0, 0.8),
+							 1px -1px 2px rgba(0, 0, 0, 0.8),
+							 -1px 1px 2px rgba(0, 0, 0, 0.8);
+}
+
+.disclaimer {
+	font-size: 0.8em;
+	opacity: 0.5;
+}
+
+.intro > button {
+	padding: 0.5em 1em;
+	margin: 2em;
+}
+
+.container > progress {
+	width: calc(100% - 2em);
+	margin: 1em;
+}
+
+.container *:focus {
+	outline: none;
+}
