+++ static/lwjgl/libraries/lwjgl.js
@@ -1,4 +1,4 @@
-var r = await fetch("/libraries/liblwjgl.so");
+var r = await fetch("/lwjgl/libraries/liblwjgl.so");
 var buf = await r.arrayBuffer();
 export default
 {
